python/requirements.txt
python-chess>=1.10.0
requests>=2.31.0
